+++ application/projects/routes.py
@@ -5,7 +5,10 @@ from flask import current_app as app
 from flask_security import roles_required, roles_accepted, current_user
 #from .assets import compile_auth_assets
 #from .forms import LoginForm, SignupForm
-from ..models import db, Project, Account
+from ..db import db
+from ..models import User, Account, Project
+from sqlalchemy import or_, and_
+from datetime import datetime
 
 
 # Blueprint Configuration
@@ -14,48 +17,97 @@ projects_bp = Blueprint('projects_bp', __name__,
                     static_folder='static')
 
 
-@projects_bp.route('/projects')
-@roles_required('super-admin')
-def projects():
-    projects = Project.query.all()
-    return render_template('projects/list.html', projects=projects)
+@projects_bp.route('/project/<id>')
+@login_required
+def project(id):
+    project = Project.query.filter_by(id=id).first()
+    return render_template('/projects/home.html', project=project)
 
+
+@projects_bp.route('/projects')
+@roles_accepted('admin','super-admin')
+@login_required
+def projects():
+    if current_user.has_role('super-admin'):
+        return redirect(url_for('manage_projects_bp.projects'))
+
+    return render_template('/projects/list.html')
 
 
 @projects_bp.route('/projects/add', methods=['POST', 'GET'])
-@roles_required('super-admin')
+@roles_accepted('admin','super-admin')
 def projects_add():
+
+    if current_user.has_role('super-admin'):
+        return redirect(url_for('manage_projects_bp.projects_add'))
+
     if 'submit-add' in request.form:
-        project = Project(name=request.form['name'], account_id=request.form['account_id'])
+        duedate = datetime.strptime(request.form['duedate'] + " 00:00:00", "%m/%d/%Y %H:%M:%S")
+        project = Project(name=request.form['name'], duedate=request.form['duedate'], account_id=current_user.account_id, creator_id=current_user.id)
         db.session.add(project)
         db.session.commit()
+        # Add Users
+        user_ids = request.form.getlist("user_ids")
+        for user_id in user_ids:
+            exists = db.session.query(User.id).filter_by(id=user_id).scalar()
+            if exists:
+                project.users.append(User.query.filter_by(id=user_id).first())
+                db.session.commit()
         return redirect(url_for('projects_bp.projects'))
-    accounts = Account.query.all()
-    return render_template('projects/form.html', template_mode='add', accounts=accounts)
+    return render_template('/projects/form.html', template_mode='add')
 
 
 @projects_bp.route('/projects/edit/<id>', methods=['POST', 'GET'])
-@roles_required('super-admin')
+@roles_accepted('admin','super-admin')
 def projects_edit(id):
-    project = Project.query.filter_by(id=id).first()
+
+    if current_user.has_role('super-admin'):
+        return redirect(url_for('manage_projects_bp.projects_edit',id=id))
+
+    project = Project.query\
+        .join(Account)\
+        .filter(and_(Project.id == id, Account.id == Project.account_id, Account.id == current_user.account_id))\
+        .first()
+
+    if not project:
+        return redirect(url_for('projects_bp.projects'))
+
     if 'submit-edit' in request.form:
         if project:
             project.name = request.form['name']
-            project.account_id = request.form['account_id']
+            project.duedate = request.form['duedate']
+            project.account_id = current_user.account_id
+            project.users[:] = []
             db.session.commit()
+            # Add Users
+            user_ids = request.form.getlist("user_ids")
+            for user_id in user_ids:
+                exists = db.session.query(User.id).filter_by(id=user_id).scalar()
+                if exists:
+                    project.users.append(User.query.filter_by(id=user_id).first())
+                    db.session.commit()
         return redirect(url_for('projects_bp.projects'))
-    accounts = Account.query.all()
-    return render_template('projects/form.html', template_mode='edit', accounts=accounts, project=project)
+    return render_template('/projects/form.html', template_mode='edit', project=project)
 
 
 @projects_bp.route('/projects/delete/<id>', methods=['POST', 'GET'])
-@roles_required('super-admin')
+@roles_accepted('admin','super-admin')
 def projects_delete(id):
-    project = Project.query.filter_by(id=id).first()
+
+    if current_user.has_role('super-admin'):
+        return redirect(url_for('manage_projects_bp.projects_delete',id=id))
+
+    project = Project.query\
+        .join(Account)\
+        .filter(and_(Project.id == id, Account.id == Project.account_id, Account.id == current_user.account_id))\
+        .first()
+
+    if not project:
+        return redirect(url_for('projects_bp.projects'))
+
     if 'submit-delete' in request.form:
         if project:
             db.session.delete(project)
             db.session.commit()
         return redirect(url_for('projects_bp.projects'))
-    accounts = Account.query.all()
-    return render_template('projects/form.html', template_mode='delete', accounts=accounts, project=project)
+    return render_template('/projects/form.html', template_mode='delete', project=project)
